+++ cgo/blockstore/ffi.go
@@ -1,4 +1,4 @@
-package cgobs
+package blockstore
 
 import (
 	"unsafe"
