+++ examples/blockstore-cgo/user/src/lib.rs
@@ -1,6 +1,6 @@
 use cid::multihash::{Code, MultihashDigest};
 use cid::Cid;
-use fvm_shared::blockstore::cgo::CgoBlockstore;
+use fvm_cgo::blockstore::CgoBlockstore;
 use fvm_shared::blockstore::{Block, Blockstore as _};
 
 #[no_mangle]
