+++ fvm/Cargo.toml
@@ -15,7 +15,7 @@ ahash = "0.7"
 num-derive = "0.3.3"
 cid = { version = "0.7", default-features = false, features = ["serde-codec"] }
 multihash = { version = "0.15", default-features = false, features = ["identity"] }
-fvm_shared = { path = "../shared", features = ["crypto", "cgo"] }
+fvm_shared = { path = "../shared", features = ["crypto"] }
 ipld_hamt = { path = "../ipld/hamt"}
 ipld_amt = { path = "../ipld/amt"}
 serde = { version = "1.0", features = ["derive"] }
